Moving the CGO functions to the header file
